# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Aura Admin API",
    "description": "TCP + bearer-token surface: config, jobs, cron, memory, traces, skills, tools, channels, LLM.",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/config": {
      "get": {
        "tags": [
          "config"
        ],
        "operationId": "get_config",
        "responses": {
          "200": {
            "description": "Current in-memory config. Secret fields are redacted by `AuraConfig`'s serde impl.",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "config"
        ],
        "operationId": "set_config",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetConfigRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Config updated on disk. Gateway restart required to pick up.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MutateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid path or value",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Write failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "config"
        ],
        "operationId": "unset_config",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnsetConfigRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Config entry removed on disk. Gateway restart required to pick up.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MutateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid path",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Write failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/v1/cron": {
      "get": {
        "tags": [
          "cron"
        ],
        "operationId": "list_cron",
        "responses": {
          "200": {
            "description": "All cron jobs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Envelope for list endpoints. Lets us add `next_cursor`, `total`,\netc. later without breaking clients that parse `items`.",
                  "required": [
                    "items"
                  ],
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "description": "Mirror of [`aura_cron::CronJob`].",
                        "required": [
                          "id",
                          "user_id",
                          "channel",
                          "schedule",
                          "action",
                          "status",
                          "created_at",
                          "updated_at"
                        ],
                        "properties": {
                          "action": {
                            "$ref": "#/components/schemas/TriggerAction"
                          },
                          "channel": {
                            "$ref": "#/components/schemas/ChannelType"
                          },
                          "created_at": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "id": {
                            "type": "string"
                          },
                          "last_triggered_at": {
                            "type": [
                              "string",
                              "null"
                            ],
                            "format": "date-time"
                          },
                          "next_trigger_at": {
                            "type": [
                              "string",
                              "null"
                            ],
                            "format": "date-time"
                          },
                          "origin_session_id": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "schedule": {
                            "$ref": "#/components/schemas/CronSchedule"
                          },
                          "status": {
                            "$ref": "#/components/schemas/CronStatus"
                          },
                          "updated_at": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "user_id": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Scheduler error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "cron"
        ],
        "operationId": "create_cron",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCronRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created cron job",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CronJob"
                }
              }
            }
          },
          "400": {
            "description": "Invalid schedule",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/v1/cron/{id}": {
      "get": {
        "tags": [
          "cron"
        ],
        "operationId": "get_cron",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Cron job id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cron job record",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CronJob"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "cron"
        ],
        "operationId": "delete_cron",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Cron job id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Cron job deleted"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Scheduler error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/jobs/{id}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "operationId": "get_job",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Job id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Job record",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApprovedResource": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "path",
              "kind"
            ],
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "read_file"
                ]
              },
              "path": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "path",
              "kind"
            ],
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "write_file"
                ]
              },
              "path": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "host",
              "kind"
            ],
            "properties": {
              "host": {
                "$ref": "#/components/schemas/HostPattern"
              },
              "kind": {
                "type": "string",
                "enum": [
                  "http"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "command",
              "kind"
            ],
            "properties": {
              "command": {
                "type": "string"
              },
              "kind": {
                "type": "string",
                "enum": [
                  "exec_command"
                ]
              }
            }
          }
        ],
        "description": "Mirror of [`aura_model::ApprovedResource`]. Paths serialize as\nstrings on the wire."
      },
      "ChannelType": {
        "type": "string",
        "description": "Admin-surface mirror of [`aura_model::ChannelType`]. Matches the\n`snake_case` wire format exactly; conversions below round-trip.",
        "enum": [
          "telegram",
          "discord",
          "http",
          "tui"
        ]
      },
      "CreateCronRequest": {
        "type": "object",
        "description": "`POST /v1/cron` body. Schedule format is the standard 5-field cron\nstring accepted by [`aura_cron`].",
        "required": [
          "schedule",
          "user_id",
          "text"
        ],
        "properties": {
          "channel": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ChannelType"
              }
            ]
          },
          "origin_session_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "schedule": {
            "type": "string"
          },
          "text": {
            "type": "string"
          },
          "user_id": {
            "type": "string"
          }
        }
      },
      "CronJob": {
        "type": "object",
        "description": "Mirror of [`aura_cron::CronJob`].",
        "required": [
          "id",
          "user_id",
          "channel",
          "schedule",
          "action",
          "status",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "action": {
            "$ref": "#/components/schemas/TriggerAction"
          },
          "channel": {
            "$ref": "#/components/schemas/ChannelType"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string"
          },
          "last_triggered_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "next_trigger_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "origin_session_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "schedule": {
            "$ref": "#/components/schemas/CronSchedule"
          },
          "status": {
            "$ref": "#/components/schemas/CronStatus"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "user_id": {
            "type": "string"
          }
        }
      },
      "CronSchedule": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "expr",
              "kind"
            ],
            "properties": {
              "expr": {
                "type": "string"
              },
              "kind": {
                "type": "string",
                "enum": [
                  "cron"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "time",
              "kind"
            ],
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "at"
                ]
              },
              "time": {
                "type": "string",
                "format": "date-time"
              }
            }
          }
        ],
        "description": "Mirror of [`aura_cron::CronSchedule`]."
      },
      "CronStatus": {
        "type": "string",
        "description": "Mirror of [`aura_cron::CronStatus`].",
        "enum": [
          "enabled",
          "disabled"
        ]
      },
      "ErrorBody": {
        "type": "object",
        "description": "Uniform error envelope for every non-2xx admin response.\n\n`GatewayError::into_response` already emits `{\"error\": \"...\"}`; this\ntype just pins the shape in the spec so clients can discriminate on\nit rather than on status codes alone.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
      "HostPattern": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "value",
              "kind"
            ],
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "exact"
                ]
              },
              "value": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "value",
              "kind"
            ],
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "wildcard"
                ]
              },
              "value": {
                "type": "string"
              }
            }
          }
        ],
        "description": "Mirror of [`aura_model::HostPattern`]."
      },
      "Job": {
        "type": "object",
        "description": "Mirror of [`aura_job::Job`].",
        "required": [
          "id",
          "session_id",
          "kind",
          "status",
          "created_at"
        ],
        "properties": {
          "completed_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "error": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string"
          },
          "input": {
            "type": [
              "object",
              "null"
            ]
          },
          "kind": {
            "$ref": "#/components/schemas/OperationKind"
          },
          "output": {
            "type": [
              "object",
              "null"
            ]
          },
          "parent_job_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "session_id": {
            "type": "string"
          },
          "started_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "trace_span_id": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "JobStatus": {
        "type": "string",
        "description": "Mirror of [`aura_job::JobStatus`].",
        "enum": [
          "pending",
          "in_progress",
          "completed",
          "submitted",
          "accepted",
          "failed",
          "stuck"
        ]
      },
...
      "MutateResponse": {
        "type": "object",
        "description": "Response body for `PUT` / `DELETE /v1/config`.",
        "required": [
          "path",
          "written_to",
          "requires_restart"
        ],
        "properties": {
          "path": {
            "type": "string"
          },
          "requires_restart": {
            "type": "boolean"
          },
          "written_to": {
            "type": "string"
          }
        }
      },
      "OperationKind": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "model",
              "type"
            ],
            "properties": {
              "model": {
                "type": "string"
              },
              "type": {
                "type": "string",
                "enum": [
                  "llm_call"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "tool_name",
              "type"
            ],
            "properties": {
              "tool_name": {
                "type": "string"
              },
              "type": {
                "type": "string",
                "enum": [
                  "tool_execution"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "skill_name",
              "type"
            ],
            "properties": {
              "skill_name": {
                "type": "string"
              },
              "type": {
                "type": "string",
                "enum": [
                  "skill_execution"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "cron_job_id",
              "type"
            ],
            "properties": {
              "cron_job_id": {
                "type": "string"
              },
              "type": {
                "type": "string",
                "enum": [
                  "cron_execution"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "strategy",
              "type"
            ],
            "properties": {
              "strategy": {
                "type": "string"
              },
              "type": {
                "type": "string",
                "enum": [
                  "context_compression"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "operation",
              "type"
            ],
            "properties": {
              "operation": {
                "type": "string"
              },
              "type": {
                "type": "string",
                "enum": [
                  "memory_operation"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "session_id",
              "type"
            ],
            "properties": {
              "session_id": {
                "type": "string"
              },
              "type": {
                "type": "string",
                "enum": [
                  "user_message_handling"
                ]
              }
            }
          }
        ],
        "description": "Mirror of [`aura_job::OperationKind`]."
      },
      "SetConfigRequest": {
        "type": "object",
        "description": "`PUT /v1/config` body.",
        "required": [
          "path",
          "value"
        ],
        "properties": {
          "path": {
            "type": "string"
          },
          "value": {
            "type": "object",
            "description": "JSON value written at `path`. Shape validated by `AuraConfig`."
          }
        }
      },
...
      "TriggerAction": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "prompt",
              "kind"
            ],
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "prompt"
                ]
              },
              "prompt": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "tool_name",
              "params",
              "approved_resources",
              "kind"
            ],
            "properties": {
              "approved_resources": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ApprovedResource"
                }
              },
              "kind": {
                "type": "string",
                "enum": [
                  "tool_call"
                ]
              },
              "params": {
                "type": "object"
              },
              "tool_name": {
                "type": "string"
              }
            }
          }
        ],
        "description": "Mirror of [`aura_cron::TriggerAction`]."
      },
      "UnsetConfigRequest": {
        "type": "object",
        "description": "`DELETE /v1/config` body.",
        "required": [
          "path"
        ],
        "properties": {
          "path": {
            "type": "string"
          }
        }
      }
    }
  },
...
}